+++ CHANGELOG.md
@@ -3,3 +3,18 @@
 ## 1.0 - Green kea
 
 - First release on PyPI of amulety
+
+## 1.1 - Blue Tui
+
+### Added
+
+- Added explicit command for BALM-paired embedding
+- Added Tutorial
+- Added better docs on included embeddings
+
+### Fixed
+
+- When multiple heavy or light chains are present, pick the one with highest duplicate count, not consensus count.
+- Consider IGH, TRB and TRD as heavy chains.
+
+### Dependencies